chore(workspace): gather tests in tests folder
