+++ app/cli/src/lib.rs
@@ -1,14 +1,5 @@
+//! Walrus CLI application.
+
 pub fn add(left: u64, right: u64) -> u64 {
     left + right
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn it_works() {
-        let result = add(2, 2);
-        assert_eq!(result, 4);
-    }
 }
