+++ app/gateway/tests/gateway.rs
@@ -1,0 +1,9 @@
+//! Tests for walrus-gateway.
+
+use walrus_gateway::add;
+
+#[test]
+fn it_works() {
+    let result = add(2, 2);
+    assert_eq!(result, 4);
+}
