+++ app/hub/tests/hub.rs
@@ -1,0 +1,9 @@
+//! Tests for walrus-hub.
+
+use walrus_hub::add;
+
+#[test]
+fn it_works() {
+    let result = add(2, 2);
+    assert_eq!(result, 4);
+}
